# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v2/minerals/projects": {
      "get": {
        "description": "Paginated, filterable catalog of mining projects across commodities (silver, gold, uranium, copper, lithium, …). Each list item includes a small `commodity_highlights` map derived from the project's resources JSONB.",
        "operationId": "MineralsController_list_v2",
        "parameters": [
          {
            "name": "commodity",
            "required": false,
            "in": "query",
            "description": "Filter by primary commodity (case-insensitive exact match).",
            "schema": {
              "example": "Silver",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by project status (case-insensitive exact match).",
            "schema": {
              "example": "Feasibility",
              "type": "string"
            }
          },
          {
            "name": "province",
            "required": false,
            "in": "query",
            "description": "Filter by province (case-insensitive exact match).",
            "schema": {
              "example": "Salta",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring search across project_name, operator, province.",
            "schema": {
              "example": "Diabl",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "project_name",
              "type": "string",
              "enum": [
                "project_name",
                "primary_commodity",
                "status",
                "province"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 50,
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProjectListItemDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationMetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta",
                    "pagination"
                  ]
                }
              }
            }
          }
        },
        "summary": "List mining projects",
        "tags": [
          "minerals"
        ]
      }
    },
    "/api/v2/minerals/projects/{name}": {
      "get": {
        "description": "Full project fact sheet, including raw `resources` and `reserves` JSONB exactly as the pipeline emitted them.",
        "operationId": "MineralsController_detail_v2",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Project name (case-sensitive, as ingested).",
            "schema": {
              "example": "Diablillos",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProjectDetailDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Project name not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Mining project detail",
        "tags": [
          "minerals"
        ]
      }
    },
...
    "/api/v2/minerals/commodities/{commodity}": {
      "get": {
        "description": "All projects whose primary commodity matches (case-insensitive), with their raw resources/reserves attached.",
        "operationId": "MineralsController_commodityProjects_v2",
        "parameters": [
          {
            "name": "commodity",
            "required": true,
            "in": "path",
            "schema": {
              "example": "Silver",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Operation",
              "type": "string"
            }
          },
          {
            "name": "province",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Salta",
              "type": "string"
            }
          },
          {
            "name": "min_resources",
            "required": false,
            "in": "query",
            "description": "If true, only include projects that have at least one resource entry.",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projects grouped under one commodity."
          }
        },
        "summary": "Projects for one commodity",
        "tags": [
          "minerals"
        ]
      }
    },
    "/api/v2/minerals/map": {
      "get": {
        "description": "Raw GeoJSON (no `data` / `meta` envelope). Projects without coordinates are excluded.",
        "operationId": "MineralsController_map_v2",
        "parameters": [
          {
            "name": "commodity",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Silver",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Feasibility",
              "type": "string"
            }
          },
          {
            "name": "province",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Salta",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "GeoJSON FeatureCollection of project points."
          }
        },
        "summary": "Projects as a GeoJSON FeatureCollection",
        "tags": [
          "minerals"
        ]
      }
    },
...
    "/api/v2/minerals/prices/{commodity}": {
      "get": {
        "description": "Latest quote for one commodity. Case-insensitive name match against the supported set.",
        "operationId": "MineralsController_price_v2",
        "parameters": [
          {
            "name": "commodity",
            "required": true,
            "in": "path",
            "schema": {
              "example": "Silver",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CommodityPriceDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Commodity not tracked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Live commodity price (single)",
        "tags": [
          "minerals"
        ]
      }
    },
...
    "/api/v1/prices/commodities/{commodity}": {
      "get": {
        "description": "Monthly value history for one commodity (`gold`, `silver`, `copper`, `lithium_etf`, `uranium_etf`).",
        "operationId": "PricesController_commoditySeries_v1",
        "parameters": [
          {
            "name": "commodity",
            "required": true,
            "in": "path",
            "schema": {
              "example": "gold",
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive start (YYYY-MM or YYYY-MM-DD).",
            "schema": {
              "example": "2025-06",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-05",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CommoditySeriesPointDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Series not tracked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Commodity time series",
        "tags": [
          "prices"
        ]
      }
    },
...
  },
  "info": {
    "title": "Petroldata.ar API",
    "description": "REST API over Argentina Secretaría de Energía oil & gas production data (focus: Vaca Muerta).\n\n## Response envelope\n\nEvery non-GeoJSON endpoint returns:\n\n```json\n{\n  \"data\": ...,\n  \"meta\": {\n    \"source\": \"Secretaría de Energía / datos.energia.gob.ar\",\n    \"dataset\": \"Producción de petróleo y gas por pozo\",\n    \"license\": \"CC-BY-4.0\",\n    \"last_source_update\": \"2026-04-01\",\n    \"last_ingested_at\": \"2026-05-27T14:30:00.000Z\",\n    \"vaca_muerta_filter\": \"formation + unconventional + sub-tipo\"\n  },\n  \"pagination\": { \"page\": 1, \"limit\": 50, \"total\": 1000 }\n}\n```\n\n`pagination` is only present on paginated list endpoints. GeoJSON endpoints (`/geo/*`) return raw GeoJSON.\n\n## Errors\n\n```json\n{ \"error\": { \"code\": \"NOT_FOUND\", \"message\": \"Well not found: 99999\" } }\n```\n\n## Units\n\n- `oil_m3` — cubic metres of oil\n- `oil_bbl` — barrels of oil (1 m³ ≈ 6.2898 bbl)\n- `oil_bbl_d` — barrels per day (oil_bbl ÷ days in month)\n- `gas_thousand_m3` — thousands of cubic metres of gas (Mm³)\n- `gas_mcf` — thousand cubic feet of gas\n- `gas_mmcf_d` — million cubic feet per day\n- `gas_mm3_d` — million m³ per day\n- `boe` — barrels of oil equivalent (oil_bbl + gas_boe)\n\n## Vaca Muerta flag\n\n`vm_combined = formation_vaca_muerta AND unconventional AND sub_resource_type ∈ {Shale, Shale Oil, …}` — pre-computed in the Python pipeline.",
...
  "servers": [
...
  "components": {
    "schemas": {
...
      "MetaDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "example": "Secretaría de Energía / datos.energia.gob.ar"
          },
          "dataset": {
            "type": "string",
            "example": "Producción de petróleo y gas por pozo"
          },
          "license": {
            "type": "string",
            "example": "CC-BY-4.0"
          },
          "last_source_update": {
            "type": "object",
            "example": "2026-04-01",
            "nullable": true
          },
          "last_ingested_at": {
            "type": "string",
            "example": "2026-05-27T14:30:00.000Z"
          },
          "vaca_muerta_filter": {
            "type": "string",
            "example": "formation + unconventional + sub-tipo"
          }
        },
        "required": [
          "source",
          "dataset",
          "license",
          "last_source_update",
          "last_ingested_at",
          "vaca_muerta_filter"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 50
          },
          "total": {
            "type": "number",
            "example": 1000
          }
        },
        "required": [
          "page",
          "limit",
          "total"
        ]
      },
...
      "ApiErrorPayloadDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "Well not found: 99999"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorPayloadDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "ProjectListItemDto": {
        "type": "object",
        "properties": {
          "project_name": {
            "type": "string",
            "example": "Diablillos"
          },
          "primary_commodity": {
            "type": "string",
            "example": "Silver"
          },
          "by_products": {
            "type": "object",
            "example": "Gold, Lead, Zinc",
            "nullable": true
          },
          "status": {
            "type": "object",
            "example": "Feasibility",
            "nullable": true
          },
          "deposit_type": {
            "type": "object",
            "example": "High Sulphidation Epithermal Style",
            "nullable": true
          },
          "province": {
            "type": "object",
            "example": "Salta",
            "nullable": true
          },
          "country": {
            "type": "object",
            "example": "Argentina",
            "nullable": true
          },
          "latitude": {
            "type": "object",
            "example": -25.3,
            "nullable": true
          },
          "longitude": {
            "type": "object",
            "example": -66.833,
            "nullable": true
          },
          "operator": {
            "type": "object",
            "example": "AbraSilver Resource Corp.",
            "nullable": true
          },
          "commodity_highlights": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "Silver": {
                "measured_kOz": 76684
              }
            },
            "description": "Per-commodity headline numbers extracted from resources/reserves for at-a-glance comparison."
          }
        },
        "required": [
          "project_name",
          "primary_commodity",
          "by_products",
          "status",
          "deposit_type",
          "province",
          "country",
          "latitude",
          "longitude",
          "operator",
          "commodity_highlights"
        ]
      },
      "CommodityPriceDto": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string",
            "example": "SI=F"
          },
          "price": {
            "type": "object",
            "example": 74.83,
            "nullable": true
          },
          "currency": {
            "type": "object",
            "example": "USD",
            "nullable": true
          },
          "previous_close": {
            "type": "object",
            "example": 74.55,
            "nullable": true
          },
          "change": {
            "type": "object",
            "example": 0.28,
            "nullable": true
          },
          "change_pct": {
            "type": "object",
            "example": 0.376,
            "nullable": true
          },
          "fifty_two_week_high": {
            "type": "object",
            "example": 121.3,
            "nullable": true
          },
          "fifty_two_week_low": {
            "type": "object",
            "example": 32.66,
            "nullable": true
          },
          "market_time": {
            "type": "object",
            "example": "2026-05-28T20:30:00.000Z",
            "nullable": true
          },
          "exchange": {
            "type": "object",
            "example": "CMX",
            "nullable": true
          },
          "instrument_type": {
            "type": "object",
            "example": "FUTURE",
            "nullable": true
          },
          "short_name": {
            "type": "object",
            "example": "Silver May 26",
            "nullable": true
          },
          "commodity": {
            "type": "string",
            "example": "Silver"
          },
          "display": {
            "type": "string",
            "example": "Silver futures (COMEX)"
          },
          "unit": {
            "type": "string",
            "example": "USD/oz"
          },
          "proxy": {
            "type": "boolean",
            "example": false,
            "description": "True when the ticker is a sector ETF proxy rather than a direct commodity feed."
          }
        },
        "required": [
          "ticker",
          "price",
          "currency",
          "previous_close",
          "change",
          "change_pct",
          "fifty_two_week_high",
          "fifty_two_week_low",
          "market_time",
          "exchange",
          "instrument_type",
          "short_name",
          "commodity",
          "display",
          "unit",
          "proxy"
        ]
      },
      "StockPriceDto": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string",
            "example": "SI=F"
          },
          "price": {
            "type": "object",
            "example": 74.83,
            "nullable": true
          },
          "currency": {
            "type": "object",
            "example": "USD",
            "nullable": true
          },
          "previous_close": {
            "type": "object",
            "example": 74.55,
            "nullable": true
          },
          "change": {
            "type": "object",
            "example": 0.28,
            "nullable": true
          },
          "change_pct": {
            "type": "object",
            "example": 0.376,
            "nullable": true
          },
          "fifty_two_week_high": {
            "type": "object",
            "example": 121.3,
            "nullable": true
          },
          "fifty_two_week_low": {
            "type": "object",
            "example": 32.66,
            "nullable": true
          },
          "market_time": {
            "type": "object",
            "example": "2026-05-28T20:30:00.000Z",
            "nullable": true
          },
          "exchange": {
            "type": "object",
            "example": "CMX",
            "nullable": true
          },
          "instrument_type": {
            "type": "object",
            "example": "FUTURE",
            "nullable": true
          },
          "short_name": {
            "type": "object",
            "example": "Silver May 26",
            "nullable": true
          },
          "operator_match": {
            "type": "string",
            "example": "AbraSilver Resource Corp."
          },
          "exchange_label": {
            "type": "string",
            "example": "OTC"
          }
        },
        "required": [
          "ticker",
          "price",
          "currency",
          "previous_close",
          "change",
          "change_pct",
          "fifty_two_week_high",
          "fifty_two_week_low",
          "market_time",
          "exchange",
          "instrument_type",
          "short_name",
          "operator_match",
          "exchange_label"
        ]
      },
      "ProjectDetailDto": {
        "type": "object",
        "properties": {
          "project_name": {
            "type": "string",
            "example": "Diablillos"
          },
          "primary_commodity": {
            "type": "string",
            "example": "Silver"
          },
          "by_products": {
            "type": "object",
            "example": "Gold, Lead, Zinc",
            "nullable": true
          },
          "status": {
            "type": "object",
            "example": "Feasibility",
            "nullable": true
          },
          "deposit_type": {
            "type": "object",
            "example": "High Sulphidation Epithermal Style",
            "nullable": true
          },
          "province": {
            "type": "object",
            "example": "Salta",
            "nullable": true
          },
          "country": {
            "type": "object",
            "example": "Argentina",
            "nullable": true
          },
          "latitude": {
            "type": "object",
            "example": -25.3,
            "nullable": true
          },
          "longitude": {
            "type": "object",
            "example": -66.833,
            "nullable": true
          },
          "operator": {
            "type": "object",
            "example": "AbraSilver Resource Corp.",
            "nullable": true
          },
          "commodity_highlights": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "Silver": {
                "measured_kOz": 76684
              }
            },
            "description": "Per-commodity headline numbers extracted from resources/reserves for at-a-glance comparison."
          },
          "by_products_list": {
            "example": [
              "Gold",
              "Lead",
              "Zinc"
            ],
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "owner_controller": {
            "type": "object",
            "example": "AbraSilver Resource Corp.",
            "nullable": true
          },
          "area_ha": {
            "type": "object",
            "example": 7919,
            "nullable": true
          },
          "technical_economic": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true,
            "example": {
              "since_production": 2029,
              "estimated_lom_years": 28,
              "productive_capacity": "9000 tpd",
              "estimated_annual_production": null,
              "capex": "620 M USD",
              "mining_method": "Open pit",
              "product": "Concentrate and doré"
            }
          },
          "resources": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true,
            "description": "Raw resources JSONB, as extracted by the pipeline."
          },
          "reserves": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true,
            "description": "Raw reserves JSONB, as extracted by the pipeline."
          },
          "resources_year": {
            "type": "object",
            "example": 2011,
            "nullable": true,
            "description": "Reporting year of the resources/reserves table (when present in the source)."
          },
          "geology": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "string"
            },
            "description": "Free-text geology sections extracted from the deck.",
            "example": {
              "regional": "The Laguna Salada Project is located near the western edge of the Cretaceous San Jorge Basin…",
              "deposit": "The mineralization at Laguna Salada is contained in flat-topped mesas…"
            }
          },
          "source_pipeline": {
            "type": "object",
            "example": "silver_2026_pptx",
            "nullable": true
          },
          "ingested_at": {
            "type": "string",
            "example": "2026-05-27T14:30:00.000Z"
          },
          "commodity_prices": {
            "description": "Live commodity prices for the project's primary commodity and any by-products we track.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommodityPriceDto"
            }
          },
          "stock": {
            "nullable": true,
            "description": "Live stock quote for the operator's parent company, if publicly traded and mapped.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/StockPriceDto"
              }
            ]
          }
        },
        "required": [
          "project_name",
          "primary_commodity",
          "by_products",
          "status",
          "deposit_type",
          "province",
          "country",
          "latitude",
          "longitude",
          "operator",
          "commodity_highlights",
          "by_products_list",
          "owner_controller",
          "area_ha",
          "technical_economic",
          "resources",
          "reserves",
          "resources_year",
          "geology",
          "source_pipeline",
          "ingested_at",
          "commodity_prices",
          "stock"
        ]
      },
...
      "CommoditySeriesPointDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2025-06"
          },
          "value": {
            "type": "number",
            "example": 3890
          }
        },
        "required": [
          "date",
          "value"
        ]
      },
...
    }
  }
}